requirements.txt
# Core RAG and LangChain dependencies
langchain>=1.2.0
langchain-google-vertexai>=3.2.0
langchain-google-community>=3.0.2
langchain-text-splitters>=1.1.0  # Required for RecursiveCharacterTextSplitter
langchain-core>=1.2.1

# PDF and document processing
unstructured[all-docs]>=0.18.0
llama-index>=0.14.0
llama-parse>=0.6.0  # High-accuracy table extraction (>99% accuracy)

# Data processing
pandas>=2.2.0
pyarrow>=22.0.0
tabulate>=0.9.0  # Required for pandas DataFrame.to_markdown()
pandas-gbq>=0.26.1  # Future requirement for BigQuery DataFrame loading (silences pandas_gbq warning)

# Google Cloud Platform
google-cloud-bigquery>=3.38.0
google-cloud-aiplatform>=1.130.0
google-cloud-storage>=2.19.0  # Required by langchain-google-vertexai

# Retrieval and ranking
flashrank>=0.2.10

# Evaluation
ragas>=0.4.1
datasets>=4.4.0

# Utilities
python-dotenv>=1.0.0

# Optional: langchain-google-genai (installed but not used - kept for reference)
# langchain-google-genai>=4.0.0  # Alternative to langchain-google-vertexai for Gemini Developer API
